Terraform modules to deploy Lambda
