+++ .gitignore
@@ -1,5 +1,14 @@
-# Created by https://www.toptal.com/developers/gitignore/api/csharp
-# Edit at https://www.toptal.com/developers/gitignore?templates=csharp
+.idea/*
+appsettings.json
+cfnstate.json
+aws-terraform/publish
+aws-terraform/lambda.zip
+aws-terraform/invoke.sh
+aws-terraform/invoke.cmd
+aws-terraform/.terraform.lock.hcl
+
+# Created by https://www.toptal.com/developers/gitignore/api/csharp,terraform
+# Edit at https://www.toptal.com/developers/gitignore?templates=csharp,terraform
 
 ### Csharp ###
 ## Ignore Visual Studio temporary files, build results, and
@@ -399,7 +408,40 @@ FodyWeavers.xsd
 # JetBrains Rider
 *.sln.iml
 
-# End of https://www.toptal.com/developers/gitignore/api/csharp
-.idea/*
-appsettings.json
-cfnstate.json
+### Terraform ###
+# Local .terraform directories
+**/.terraform/*
+
+# .tfstate files
+*.tfstate
+*.tfstate.*
+
+# Crash log files
+crash.log
+crash.*.log
+
+# Exclude all .tfvars files, which are likely to contain sensitive data, such as
+# password, private keys, and other secrets. These should not be part of version
+# control as they are data points which are potentially sensitive and subject
+# to change depending on the environment.
+*.tfvars
+*.tfvars.json
+
+# Ignore override files as they are usually used to override resources locally and so
+# are not checked in
+override.tf
+override.tf.json
+*_override.tf
+*_override.tf.json
+
+# Include override files you do wish to add to version control using negated pattern
+# !example_override.tf
+
+# Include tfplan files to ignore the plan output of command: terraform plan -out=tfplan
+# example: *tfplan*
+
+# Ignore CLI configuration files
+.terraformrc
+terraform.rc
+
+# End of https://www.toptal.com/developers/gitignore/api/csharp,terraform
