+++ aws-terraform/cloudwatch/main.tf
@@ -1,0 +1,4 @@
+resource "aws_cloudwatch_log_group" "log_group" {
+  name              = "/aws/lambda/${var.lambda_name}"
+  retention_in_days = 14
+}
