+++ aws-terraform/cloudwatch/variables.tf
@@ -1,0 +1,4 @@
+variable "lambda_name" {
+  description = "The name of the Lambda function"
+  type        = string
+}
