+++ aws-terraform/dynamodb/variables.tf
@@ -1,0 +1,4 @@
+variable "id_prefix" {
+  description = "The prefix to IDs/names of AWS resources"
+  type        = string
+}
