+++ aws-terraform/locals.tf
@@ -1,0 +1,3 @@
+locals {
+  lambda_name = "${var.id_prefix}-lambda"
+}
