+++ aws-terraform/s3/variables.tf
@@ -1,0 +1,9 @@
+variable "id_prefix" {
+  description = "The prefix to IDs/names of AWS resources"
+  type        = string
+}
+
+variable "appsettings_source" {
+  description = "The source path of the appsettings.json file"
+  type        = string
+}
