+++ aws-terraform/scheduler/main.tf
@@ -1,0 +1,53 @@
+resource "aws_scheduler_schedule" "schedule" {
+  name       = "${var.id_prefix}-schedule"
+  group_name = "default"
+
+  flexible_time_window {
+    mode = "OFF"
+  }
+
+  schedule_expression = var.schedule_expression
+
+  target {
+    arn      = var.lambda_arn
+    role_arn = aws_iam_role.scheduler_role.arn
+  }
+}
+
+resource "aws_iam_role" "scheduler_role" {
+  name = "${var.id_prefix}-scheduler-role"
+
+  assume_role_policy = jsonencode({
+    Version = "2012-10-17"
+    Statement = [
+      {
+        Action = "sts:AssumeRole"
+        Effect = "Allow"
+        Principal = {
+          Service = "scheduler.amazonaws.com"
+        }
+      }
+    ]
+  })
+}
+
+resource "aws_iam_policy" "scheduler_invoke_policy" {
+  name        = "${var.id_prefix}-scheduler-policy"
+  description = "IAM policy for EventBridge Scheduler to invoke Lambda function"
+
+  policy = jsonencode({
+    Version = "2012-10-17"
+    Statement = [
+      {
+        Action   = "lambda:InvokeFunction"
+        Effect   = "Allow"
+        Resource = [var.lambda_arn]
+      }
+    ]
+  })
+}
+
+resource "aws_iam_role_policy_attachment" "scheduler_invoke_policy_attachment" {
+  role       = aws_iam_role.scheduler_role.name
+  policy_arn = aws_iam_policy.scheduler_invoke_policy.arn
+}
